pyproject.toml
[project]
name = "aiengdict"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "commitizen>=4.9.1",
    "pre-commit>=4.3.0",
    "pyclean>=3.1.0",
    "pytest>=8.4.2",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.0.0",
    "pytest-env>=1.1.5",
    "flask>=3.1.2",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-login>=0.6.3",
    "sqlalchemy>=2.0.36",
]
